+++ src/test/suite/Blame.test.ts
@@ -3,6 +3,7 @@ import * as mocha from 'mocha';
 import * as sinon from 'sinon';
 import * as blameMock from '../../Blame';
 import Notifications from '../../Notifications';
+import Settings from '../../Settings';
 import { activateExtension, document } from './helpers';
 
 export const getBlame = () => `14728a3afb0871ac18e84d4d3e41bd34f0283170 1 1 3
@@ -42,6 +43,13 @@ filename src/test.ts
         // foobar
 14728a3afb0871ac18e84d4d3e41bd34f0283175 2 6
 author &%¤#"!?=(){}[]/\\
+author-mail <test@example.com>
+author-time 1743513058
+summary This is a test commit message
+filename src/test.ts
+        // foobar
+14728a3afb0871ac18e84d4d3e41bd34f0283176 2 7
+author Testing Tester TT
 author-mail <test@example.com>
 author-time 1743513058
 summary This is a test commit message
@@ -114,13 +122,14 @@ suite('Test Blame', () => {
     });
 
     test('test blame function', async () => {
+        const settings = sinon.stub(Settings, 'getAuthorStyle').returns('full');
         mock.expects('blameFile').returns(getBlame());
 
         const blamed = await blameMock.blame(document);
 
         const first = blamed[0];
         
-        assert.strictEqual(first.author, 'Jami Lu');
+        assert.strictEqual(first.author.displayName, 'Jami Lu');
         assert.strictEqual(first.linenumber, '1');
         assert.strictEqual(first.email, 'test@example.com');
         assert.strictEqual(first.summary, 'This is a test commit message');
@@ -128,12 +137,51 @@ suite('Test Blame', () => {
         assert.strictEqual(first.date.dateString, '2024417');
         assert.strictEqual(first.filename, 'src/test.ts');
         assert.ok(first.codeline.includes('import * as vscode from "vscode"'));
-        assert.strictEqual(blamed[1].author, 'über D.ber');
+        assert.strictEqual(blamed[1].author.displayName, 'über D.ber');
         assert.ok(blamed[1].codeline.includes('// foobar'));
-        assert.strictEqual(blamed[2].author, '李 连杰');
-        assert.strictEqual(blamed[3].author, 'שלום שלום');
-        assert.strictEqual(blamed[4].author, 'r/a\\b test');
-        assert.strictEqual(blamed[5].author, '&%¤#"!?=(){}[]/\\');
-        assert.strictEqual(6, blamed.length);
+        assert.strictEqual(blamed[2].author.displayName, '李 连杰');
+        assert.strictEqual(blamed[3].author.displayName, 'שלום שלום');
+        assert.strictEqual(blamed[4].author.displayName, 'r/a\\b test');
+        assert.strictEqual(blamed[5].author.displayName, '&%¤#"!?=(){}[]/\\');
+        assert.strictEqual(blamed[6].author.displayName, 'Testing Tester TT');
+        assert.strictEqual(blamed.length, 7);
+
+        settings.restore();
+    });
+
+    test('test blame annotation first name', async () => {
+        const settings = sinon.stub(Settings, 'getAuthorStyle').returns('first');
+        mock.expects('blameFile').returns(getBlame());
+
+        const blamed = await blameMock.blame(document);
+
+        assert.strictEqual(blamed[0].author.displayName, 'Jami');
+        assert.strictEqual(blamed[1].author.displayName, 'über');
+        assert.strictEqual(blamed[2].author.displayName, '李');
+        assert.strictEqual(blamed[3].author.displayName, 'שלום');
+        assert.strictEqual(blamed[4].author.displayName, 'r/a\\b');
+        assert.strictEqual(blamed[5].author.displayName, '&%¤#"!?=(){}[]/\\');
+        assert.strictEqual(blamed[6].author.displayName, 'Testing Tester');
+        assert.strictEqual(blamed.length, 7);
+
+        settings.restore();
+    });
+
+    test('test blame annotation last name', async () => {
+        const settings = sinon.stub(Settings, 'getAuthorStyle').returns('last');
+        mock.expects('blameFile').returns(getBlame());
+
+        const blamed = await blameMock.blame(document);
+
+        assert.strictEqual(blamed[0].author.displayName, 'Lu');
+        assert.strictEqual(blamed[1].author.displayName, 'D.ber');
+        assert.strictEqual(blamed[2].author.displayName, '连杰');
+        assert.strictEqual(blamed[3].author.displayName, 'שלום');
+        assert.strictEqual(blamed[4].author.displayName, 'test');
+        assert.strictEqual(blamed[5].author.displayName, '&%¤#"!?=(){}[]/\\');
+        assert.strictEqual(blamed[6].author.displayName, 'TT');
+        assert.strictEqual(blamed.length, 7);
+
+        settings.restore();
     });
 });